refactor(comp): Change dir component format to Component.tsx
