+++ src/home/pages/react/exProps/ExProps.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import React from 'react';
 import {DynamicPropsFc} from './dynamicProps/DynamicPropsFc';
 import DynamicPropsClass from './dynamicProps/DynamicPropsClass';
-import RequiredProps from './requiredProps/RequiredProps';
+import RequiredProps from './RequiredProps';
 import {Highlighter} from '@common/highlighter';
 import {
     RequiredPropsCode,
