+++ src/home/pages/react/exReduxTodo/ExReduxTodo.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 import React from 'react';
-import {TodoList} from './todoList/TodoList';
-import {TodoEditor} from './todoEditor/TodoEditor';
+import {TodoList} from './TodoList';
+import {TodoEditor} from './TodoEditor';
 
 export const ExReduxTodo = () => {
