+++ src/home/pages/react/index.ts
@@ -1,10 +1,10 @@
-import {Welcome} from './welcome/Welcome';
+import {Welcome} from './Welcome';
 import {ExState} from './ExState';
 import {ExProps} from './exProps/ExProps';
 import {ExTodoList} from './exTodoList/ExTodoList';
 import {ExRedux} from './exRedux/ExRedux';
 import {ExReduxTodo} from './exReduxTodo/ExReduxTodo';
-import ExLifeCycle from './exLifeCycle/ExLifeCycle';
+import ExLifeCycle from './ExLifeCycle';
 import {ExBasicSyntax} from './exBasicSyntax/ExBasicSyntax';
 import {ExCart} from './exCart/ExCart';
 import {ExVeneer} from './exVeneer/ExVeneer';
